requirements.txt
Flask==3.1.0
Flask-Session==0.8.0
mysql-connector-python==9.3.0
Pillow==10.4.0
python-dotenv==1.0.1
Werkzeug==3.1.3
